requirements.txt
pycryptodome=3.9.0
certifi==2019.9.11
pipenv==2018.11.26
pybat==0.0.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
